+++ modules/hyprland.nix
@@ -12,7 +12,6 @@
   programs.hyprland.xwayland.enable = true;
 
   programs.hyprlock.enable = true;
-  services.hypridle.enable = true;
 
   # Force electron apps to use wayland instead of X11.
   environment.sessionVariables.NIXOS_OZONE_WL = "1";