+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 In this file I'll try to mention major changes to the configuration.
 
 2025-11-25
-- hyprland: added screenshot functionalities and hotkeys for them
+- hyprland: added screenshot functionalities and hotkeys for them.
+- hyprland: configure auto lock screen actions.
 
 2025-11-24
 - hyprland: added packages that help with file thumbnails.